refact(hooks/useTableLogic, utils/navigation): componentizando lógica de ordenação, busca e rolagem
