+++ src/utils/navigation.jsx
@@ -1,0 +1,11 @@
+export const goRegistration = (navigate, routeName) => {
+  navigate(`/${routeName}/novo`);
+};
+
+export const goView = (navigate, routeName, id) => {
+  navigate(`/${routeName}/${id}`);
+};
+
+export const goEdit = (navigate, routeName, id) => {
+  navigate(`/${routeName}/${id}/editar`);
+};
